chore: GSE docs and renames (#15810)
Adds clarifying comments/renames primarily to GSE and AddressSnapshotLib

The big change is being very explicit about what "canonical" means,
which is reserved for "the latest rollup according to the registry".

The GSE has its own, separate view of "rollups" that have been added to
it.

"Rollups" are "instances" according to the GSE, and there is a special
"bonus" instance that is available for use by the most recently added
rollup.