chore: remove public_inputs from DeciderVK (#15832)
Removes explicit storage of `public_inputs` in the DeciderVK. The public
inputs are folded as part of the wire polynomials, not as individual
elements so its not meaningful for the verifier accumulator to store
them explicitly (which has recently led to some confusion in IVC state
mocking). They are now stored directly in the oink/PG verifiers where
they are populated from the proof and used to reconstruct the IO data.