chore: delete LegacyPolynomial (#15847)
`LegacyPolynomial` is only used in `polynomial.test.cpp`, where we can
use `Polynomial` instead. Some test cases were deleted because there was
no corresponding method to test in `Polynomial`:
- `factor_roots`: this tests multiple roots but `Polynomial` only has a
method for a single root.
- `RightShift`: There's no similar method for `Polynomial`.
`right_shifted` works differently and points to a different block of
memory of the same size.
- `evaluate_mles`: it compares results between `Polynomial` and
`LegacyPolynomial` so makes no sense when `LegacyPolynomial` is deleted.