chore: remove circuit merkle membership. (#15794)
Removes the circuit implementation for merkle membership checks in cpp.
This implementation isn't being used anywhere (including noir) and we
decided to remove it to avoid maintenance (and audit).

As a result, corresponding circuit benchmarks for merkle membership
checks is removed. I also had to adjust the parameter (number of
instantiations) for mock SHA-256 and ECDSA to ensure we still use 2^17
and 2^19 as medium and large mock circuits for CIVC testing.