chore: Update `UltraVerifier` api to return public inputs in the case of a MegaFlavor (#15868)
After
[#15382](https://github.com/AztecProtocol/aztec-packages/pull/15832),
Decider VKs do not contain public inputs anymore. In
[#15829](https://github.com/AztecProtocol/aztec-packages/pull/15829) we
need to access the public inputs to reconstruct the last merged table.
This PR moves the reconstruction inside the UltraVerifier verify
function and define two `verify` in two: one for `UltraFlavor` (which
returns `bool`) and one for `MegaFlavor` (which returns `bool` and the
last merged table).

Closes issue
[#1483](https://github.com/AztecProtocol/barretenberg/issues/1483).

---------

Co-authored-by: AztecBot <tech@aztecprotocol.com>
Co-authored-by: ledwards2225 <98505400+ledwards2225@users.noreply.github.com>