fix: Propose before voting in multicall (#15920) (#15927)
Backport for #15920

Attempt at fixing the OOG we've been seeing on adversarial testnet when
sequencers try to vote and propose at the same time.

- Reenables the gov-proposal test, which tested this case and had been
disabled for months.
- Bumps the number of validators in the test to 48 (all in the same node
though) and includes outbox msgs to try and reproduce the oog. I failed
though: I had to manually tweak the `VOTE_GAS_GUESS` to ridiculously low
values to reproduce the oog.
- Changes the sequencer publisher so it always calls `propose` first, so
the committee is pre-stored in transient storage. It's unclear to me if
this will still be the case after all the work in delayed signatures.
- Bumps the vote gas guess to 800k, since we've seen that votes were
taking 700k gas in sepolia.