chore: remove large swaths of made-deprecated txe code (#15934)
We're almost there. This clears the last stragglers that prevented
removing `private()` and friends, removes unused files, tests, helpers.
The `TestEnvironment` impl is _almost_ sane after this PR. Some
polishing is required on naming, devex, and I need to take a look at
deployment, but the testing front-end is _almost_ there.

Built on top of
https://github.com/AztecProtocol/aztec-packages/pull/15928, merge that
first.

Closes https://github.com/AztecProtocol/aztec-packages/issues/8588.