chore: remove call machinery from txe  (#15935)
And here we finally delete all of the machinery required to process
private and public calls inline, (i.e. the old
`call_iface.call(context)`, which we don't do anymore. I am not entirely
sure if this will pass CI in it's current state, but detecting usage of
this feature is a bit hard unfortunately so this is a good way of seeing
what fails.

Built on top of
https://github.com/AztecProtocol/aztec-packages/pull/15934, merge that
first.