chore: Modify the `MergeVerifier` so that it gets the subtable commitments as input and returns the commitment to the merged table (#15949)
We modify the `MergeVerifier` so that it gets the subtable commitments as input and returns the commitment to the merged table. The reason for this change is that given the new structure of `ClientIVC` following [#15704](https://github.com/AztecProtocol/aztec-packages/pull/15704), we can't access the merged table commitments from inside `complete_hiding_circuit_logic`.

This PR is in preparation for [#15829](https://github.com/AztecProtocol/aztec-packages/pull/15829)

---------

Co-authored-by: AztecBot <tech@aztecprotocol.com>