feat: Add the last merged table to the public inputs of the hiding circuit (#15829)
We make the merged table received by the Merge verifier in the hiding
circuit a public input to the hiding circuit. This is needed because the
Merge verifier will soon receive `t_commitments`, `T_prev_commitments`
as inputs rather than reading them from the proof.

**EDIT:**

To complete the work on the consistency checks, and to ensure the
soundness of the Goblin verification, the merged table received by the
Merge verifier in the last step of a Goblin accumulation must be set to
be a public input of the circuit that performs the verification, so that
the verifier can extract that public input and use it as the commitment
to the previous table in the Merge verification.

For example, in ClientIVC the last Merge verification before the final
Goblin verification happens in the HidingKernel, so we need to add the
merged table commitments received by the Merge verifier inside the
HidingKernel to be public inputs of the HidingKernel.

After this PR, `MegaVerifier = UltraVerifier<MegaFlavor>` always expects
the inputs to be `PairingInputs` + commitments to ECC op tables. These
inputs are produced by the class `HidingKernelIO` (even though in the
future we might consider changing this name)

The PR required changes to various tests to accommodate the new
structure of the public inputs.

---------

Co-authored-by: AztecBot <tech@aztecprotocol.com>
Co-authored-by: ledwards2225 <98505400+ledwards2225@users.noreply.github.com>
Co-authored-by: sergei iakovenko <105737703+iakovenkos@users.noreply.github.com>
Co-authored-by: ludamad <adam.domurad@gmail.com>
Co-authored-by: Suyash Bagad <suyash@aztecprotocol.com>
Co-authored-by: Jonathan Hao <jonathan@aztec-labs.com>
Co-authored-by: Raju Krishnamoorthy <krishnamoorthy@gmail.com>
Co-authored-by: notnotraju <raju@aztec-labs.com>
Co-authored-by: Lucas Xia <lucasxia01@gmail.com>
Co-authored-by: Khashayar Barooti <khashayar@aztecprotocol.com>
Co-authored-by: Jean M <132435771+jeanmon@users.noreply.github.com>
Co-authored-by: Alex Gherghisan <alexghr@users.noreply.github.com>
Co-authored-by: Santiago Palladino <spalladino@users.noreply.github.com>
Co-authored-by: Santiago Palladino <santiago@aztec-labs.com>
Co-authored-by: ludamad <domuradical@gmail.com>
Co-authored-by: maramihali <mara@aztecprotocol.com>
Co-authored-by: Sarkoxed <75146596+Sarkoxed@users.noreply.github.com>