fix: wonky out hash tree (#16073)
This pr fixes the function that constructs the membership witness for
consuming the l2ToL1Msg.

It also changes the circuits to build a real wonky tree for the tx out
hash.
Previously it was a variable tree, which had 0s padded to the shorter
branch to make it a balanced tree.
Changing it to a wonky tree doesn't make it cheaper to consume the msg
on L1, because the leaf in the shorter path still has the same height.
But it does make it cleaner to use the same code to compute the out
hashes in ts, for both tx level and block level.

---------

Co-authored-by: Michael Connor <mike@aztecprotocol.com>