feat: specialised ZeroSelector to save memory on selectors known to be zero (#16141)
This is a reland of
https://github.com/AztecProtocol/aztec-packages/pull/16044 with a minor
fix
https://github.com/AztecProtocol/aztec-packages/pull/16141/commits/7e9b998c7e2ad94f6d6aab7fb161d8766ea6c633.

Original description:

Most ExecutionTraceBlock except the overflow block is restricted to a
certain type of gates. The other gate selectors are always zero so we
don't need to store them explicitly. Not storing them can save us about
400MB memory.

I felt it was the easiest to represent the selectors that will be zero
by a new class of selector ZeroSelector, which is always zero and uses
constant memory. This way, the call sites that populate or read the
selectors only need minimum change.

Each selector in the block is of an abstract type Selector, and they can
be either ZeroSelector or SlabVectorSelector depending on which derived
class the blocks belong to.

For the bigger picture, see https://hackmd.io/7wcdcdLtQGy95qg7xMFH9A.

Benchmarks:

Testing the memory limit by running CPUS=16 MEM=Xg
~/aztec-packages/ci3/docker_isolate "BB_SLOW_LOW_MEMORY=1
./build/bin/ultra_honk_bench
--benchmark_filter='construct_proof_ultrahonk_power_of_2/21$'" with
gradually smaller X. On this branch, X can be as low as ~1.3, whereas on
the merge-train, X can only go as low as ~1.7. This confirms the 400MB
memory improvement.