feat!: update breaking changes for aztec contract testing in noir in docsite (#16118)
This adds migration notes for the changes to noir testing i've
introduced over the past couple weeks, and updates all mentions of txe
from the docsite to show the new API and naming.

Note that this does not improve the current docs for testing - it just
replaces the old terms with new ones, and removes things that should not
be exposed. I took the liberty of adjusting some of the example code
samples were the code we were displaying was too jarring. I may do a
pass in the near future going over the state of the API, sample tests,
and the way to use calls and contexts.

As of this PR, the name 'TXE' goes away from userland in favor of the
broader concept of 'testing aztec contracts in noir', or the
`TestEnvironment` module. TXE remains as an implementation detail, but
does not need to be exposed to users.

---------

Co-authored-by: josh crites <critesjosh@gmail.com>