feat: Validators invalidate invalid blocks (#16120)
We expect proposers to invalidate the previous block if it is invalid,
but if they fail to do so, validators will eventually do it,
prioritizing the committee members and then any validator whatsoever.

This commit includes other fixes:
- If a proposer cannot build a block due to not enough txs, it still
tries to invalidate the previous one.
- The archiver keeps track of the earliest (not latest) invalid block it
has seen, so the sequencer can use this info to invalidate the earliest
one.

Builds on top of #16067