feat: remove duplication of hiding circuit creation by handling its completion outside of ClientIVC::prove (#16135)
This PR changes the way the construction of the hiding kernel is handled
to accommodate for the hiding kernel being implemented in Noir.
Changes include: 
- added two methods `construct_and_accumulate_next_circuit` and
`construct_hiding_circuit` to
`PrivateFunctionExecutionMockCircuitProducer`, to handle creation of the
hiding circuit in our testing modules
- removed the duplication of `complete_hiding_circuit_logic`. Now the
`hiding_circuit` builder is saved as a member `ClientIVC` class.
- added a `construct_and_accumulate_mock_kernel` method to
`ivc_recursion_constraints` test module.
- updated the API according to the changes to the `ClientIVC` class.

---------
Co-authored-by: Khashayar Barooti <khashayar@aztecprotocol.com>