chore: Template ultra verifier `verify_proof` over the public inputs type (#16239)
The `UltraVerifier` and `UltraRecursiveVerifier` choose which public
inputs to extract based on the `Flavor` with which they are
instantiated. This is not advisable as it forces us to change testing
code (so that `Mega` circuits have inputs of the form `HidingKernelIO`)
and forces us to pay for unused public inputs when `Mega` circuits don't
have ecc op tables (e.g., the `GoblinAvmRecursiveVerifier`).

This PR changes the `verify_proof` function of both verifiers. We
template the function over the `IO` class, which specifies which public
inputs should be extracted from the proof. We make native and recursive
verifiers uniform by modifying the return type of `verify_proof` in
native land: we change it from `bool` to `UltraVerifierOutput`, which is
a new struct that contains a boolean (the result of the verification)
and ecc op tables (which are only filled when `UltraVerifier` is used to
verify a `ClientIVC` proof).

We also remove the `verify_proof` function in the recursive verifier
that took a `HonkProof` instead of an `Stdlib::Proof`. The reason is
that we to specialise `verify_proof` over the various `IO` types, and
the compiler was complaining about ambiguous specialisations.

This PR closes
https://github.com/AztecProtocol/barretenberg/issues/1489.

---------

Co-authored-by: AztecBot <tech@aztecprotocol.com>
Co-authored-by: ludamad <adam.domurad@gmail.com>