feat: mechanism to add randomness to the op queue (#16221)
Added a random op type to the EccOpQueue class and a mechanism in the
`MegaCircuitBuilder` to add such an operation to an op queue (all ops
are added in a realistic scenario through the builder). Also added
testing to ensure that if we produce a MegaHonk proof for a circuit
whose ecc ops table contain randomness, this would not affect the
correctness of the MegaHonk proof. The mechanism is not yet used in
CIVC.

---------
Co-authored-by: ledwards2225 <l.edwards.d@gmail.com>