feat: merge via append the ecc ops in hiding kernel (#16267)
Modified the logic in CIVC class and related tests to support merging
via append the subtable of ecc ops coming from the hiding kernel to the
total table. With this modification, all CIVC tests have to accumulate a
meaningful number of circuit to reach a tail kernel (which recursively
verifies a proof labelled as `PG_TAIL`). This is due to the fact that it
is the tail kernel that will now start with a no-op in order to ensure
that we can correctly shift the polynomials representing the op queue in
translator.