feat: replace bb_cli_bench and op-count-time build with bb and runtime flag (#16189)
`bb` takes --print_benchmark or --benchmark_out=<json>. If either is
given, it will run with op-count-time in a benchmark.
`--print_benchmark` only prints the total time but not the op-count-time
though, I'm not sure how to get that from Google benchmark.

Other binaries can be run with op-count-time with env
`BB_USE_OP_COUNT_TIME=1`.

There's still a 1% overhead even when the env or the runtime flag is
disabled because we need to check the flag anyway.

Fixes https://github.com/AztecProtocol/barretenberg/issues/1458

---------

Co-authored-by: AztecBot <tech@aztecprotocol.com>
Co-authored-by: maramihali <mara@aztec-labs.com>
Co-authored-by: ludamad <domuradical@gmail.com>