chore(ci): experiment with non-linear history
⚠[aztec-packages](https://github.com/AztecProtocol/aztec-packages) Developer Notice⚠
TL;DR: We are experimenting with a move away from strict squashing in merge queue. See this PR. If you *aren't targeting a merge-train*, you likely should use the label `ci-squash-and-merge` now.
If your team wants a merge-train branch, talk to Adam.

**Rationale:**
Merge-train is a concept we are experimenting with to allow less strict merges to a team branch through a feature branch with automation to merge it into next and vice-versa.
Because of merge-train, we want to allow a mix of squashes and merges into merge queue. This will make feature branches much more viable as a means to bundle changes before sending to merge queue as their history is not lost due to a squash, and developers who target a merging merge-train will not require tedious rebasing. This means however, because of the lack of customizability of merge queue, that normal PRs could merge with merge commits too.

**Developer Workflow Changes:**
**There is now a check enforcing that PRs into `next` have only one commit.* If you are ready to merge you need to push a squash commit. One shortcut for this is the label `ci-squash-and-merge` which will squash your PR and put it in the merge queue after it passes ci3.yml build/tests. Generally, the hope is you can switch to that instead of the "Merge when Ready" button and move on. Also, the tradeoffs of merge-train should be a lot more attractive now.
If you want to merge your PR history as-is you can now use ci-no-squash , but there should rarely be need for this.