feat(docs): Add docusaurus plugin to autogerate llms.txt
Adds a plugin to the docs to autogenerate an llms.txt file and an llms-full.txt file to be published with the docs site according to the standards defined at https://llmstxt.org/.

Also adds page descriptions to pages that are missing them, since this helps with generating the llms.txt.

Note that this is set up to generate the file for the first versioned defined in the versions.json file, which should be the default version for the docs site.

Also removes an outdated sandbox features page.