feat(avm): separate accumulate instantiations to own cpp
This PR separates all template instantiations of `Relation::accummulate`​ to their own cpp file. This improves paralellizability and removes the load on files like `prover.cpp`​, `verifier.cpp`​, etc.

For relations

- `generated/relations/{{relation}}.hpp`​ has only the declaration of `accumulate`​.
- `generated/relations/{{relation}}_impl.hpp` has only the definition of `accumulate`​.
- `generated/relations/{{relation}}.cpp`​ has (almost all) the instantiations of `accumulate`​.
- `generated/relations/relation_impls.hpp` includes all impls (used only in benches and possibly tests).

For lookups

- `constraining/relations/interactions_base.hpp` has only the declaration.
- `constraining/relations/interactions_base_impl.hpp` has only the definition.
- `generated/relations/lookups_{{relation}}.cpp`​ instantiates the definition.

For permutations (to be done, I have to mirror what I did for lookups).

Results

- ARM compilation time goes down from 21m to 10m (see note later)
- x86 time gets reduced by ~1m
- Memory during compilation now peaks briefly at 58GB, but otherwise mostly doesnt use much memory.

For ARM, now the poseidon perm cpp becomes the bottleneck, but this will be solved by Ilyas’ PR. So the time will go down to 9m. If we find some other improvements for the recursive verifier, we might be able to take it down to 5m and ~parity with x86.

![image.png](https://app.graphite.dev/user-attachments/assets/61c88e85-8487-42ae-b586-62a76806bb8e.png)