feat!: add test env note discovery
This adds `TestEnvironment::discover_note`, which is needed to emulate the note discovery that happens in real contracts for notes created in a `TestEnvironment::pirvate_context`. With this we can now extend our testing of note-based state variables, as well as properly tests constructs such as `DelayedPrivateMutable`.

I introduced some minor adjustments to the internals of the message encoding functions to better reuse those capabitilies. These are part of a larger set of changes we'll likely want to introduce as we unify and improve the note and event emission APIs.

edit: I also added some tests for note destruction and checking that we do not return those notes, both in transient and settled scenarios. I'm only testing private context and not utility context to keep the number of tests here small, but we should later expand these, perhaps with some more organization of the internals of the note getter API.

Interestingly, I had to pseudo introduce the notion of 'syncing' and how it relates to nullified note discovery (#12553) here, since  direct context usage does not run the autogenerated note discovery code. This resulted in some findings, notably the note data provider being a bit annoying about requiring knowledge of the recipient of the notes that were being nullified, seemingly for no good reason. I had to remove all of that because `env::discover_note` sets the zero address as the recipient as none are assumed to exist, and so when looking up notes we'd want to check for nullification we'd find none since our PXE had no accounts. This now searches _all_ notes in scope, regardless of who their recipient is.