test: AMM TXE tests
Fixes #10289

Just like my [other recent PR](https://github.com/AztecProtocol/aztec-packages/pull/16451) decided to add TXE tests for AMM.

## On using AI here
The only thing I did manually here is to merge a few test cases because they were too long-running as setting up the initial state is costly. I also told it to not have precise ratio in the second call to `add_liquidity` to test the refund. Other than this AI pretty much managed to nail it.

### Update on AI
Code quality was not great and had to redo quite a lot of stuff. That could be improved though if we had a proper reference (plenty of the style issues were present in the rest of our codebase).