chore: Validate L1 config in ts-land
Validates L1 config _before_ trying to deploy L1 contracts, so we can
get an early warning if the deployment would fail.

Also, changes the slashRoundSize setting to slashRoundSizeInEpochs, to
force it to be a multiple of epochs, as required by the slash proposer.
It also makes the quorum settings optional, defaulting them to half the
round size plus one. This should make configuration easier.