fix: Update slash settings in helm templates
Updates the slasher variables used in helm templates to match the new
ones defined in #16694 and #16625. Sets all values to be empty, so we
rely on the defaults set in the node and don't have multiple places
where we define default values.

Also adds a `check_env_vars` script (authored by claude) that checks if
we are using any env var not defined in the env_var list in ts, so the
CI should shout if we update a variable in ts-land but forget to update
it in helm.