feat: oracle version check
Fixes #14979

This PR implements a version compatibility system between Aztec.nr contracts and the PXE oracle interface to prevent ambiguous errors from version mismatches. The oracle version is expected to be manually updated but to ensure that this does not get forgotten the build throws an error if an oracle interface hash constant needs updating.