feat!: merge the txe session state transition oracles, redo the internal wiring, introduce context handlers
This one leaves things in a bit of a messy state, as doing _the entire_ transition is unfeasiable, but the things that are ready _do_ look good, and it's a good starting place for future changes.

The session now calls the translator (service) with a handler for state changes (the session itself) and a handler for all other calls. On session state transitions, a new handler is installed and the old one closed and committed to. There's now a handler for `public_context` which does the avm opcodes, and consequently the `txe_oracle` no longer knows how to process these (because it doesn't need to).

I also cleaned up the state transition oracles to take all params in one go instead of relying on mutating global session state, which helps make this smoother and also helps with the cleanup and eventual removal of the `txe_oracle`.

Co-authored-by: Nicolás Venturo <nicolas.venturo@gmail.com>