chore: `stdlib::poseidon2` internal audit (#16534)
- Most of the `Poseidon2`-related methods and classes were unnecessarily
templated on a variety of parameters, such as `rate` and `capacity` of
the permutation, while at the same time our Circuit Builders would only
support the case `rate = 3`, `capacity = 1`. To avoid confusion and
improve the readability, I minimized the number of templates.
- The `sponge` class was only invoked with a sequence of `absorb` calls
that would lead to a single `squezze` call. It allowed to simplify the
class logic.
- Simplified the `Poseidon2Permutation` class by re-using some of the
`field_t` methods. Improved the docs.
- Added a detailed **readme** that contains links to the sources,
including the padding choices in `sponge` and the explanation of the
custom Poseidon2 gates
- Audited relations, expanded corresponding docs
- Added a bunch of tests - collision-resistance sanity check, test
against the hash outputs from https://github.com/zemse/poseidon2-evm,
several basic failure tests