chore(avm)!: reverting pr 16856 (#16888)
The PR #16856 (migration of gas related lookups to GT gadget) caused a
major performance regression for the commitment computation of the
logDerivativeInverses in the bulk trace.

We suspect that the number of inverses in the gt gadget trace strongly
increased. As gas is involved in every opcode, the impact was
significant.