feat: `utilityGetUtilityContext` oracle
Introduces a new `utilityGetUtilityContext` oracle that returns all the info needed for `UtilityContext` construction in a single call. This reduces the number of `AztecNode` calls from 5 to 1. This should significantly speed up utility function execution when there is a high latency between PXE and the node.

## Bug fix
Before this PR there was a serious bug in the `UtilityContext` construction because it was constructed based on values obtained from aztec node. This is just wrong because the simulation should be run against the `PXE`'s view of the network state and not against the node's view. I think we didn't stumble upon this being an issue because we perform a sync right when the simulation is started. But I think it would be inevitable to happen that a node would sync a block from p2p in between the initial `PXE` sync and construction of the utility context and at that point we would enter an undefined behavior and a world of pain of debugging.

## Followup work
As Nico mentioned in review round 1 seems to make sense to nuke the following oracles:
- `utilityGetChainId`
- `utilityGetVersion`
- `utilityGetTimestamp`

and possibly also
- `utilityGetBlockNumber`

as all these values are contained inside the response from the `utilityGetUtilityContext` oracle. That would also allow me to nuke these handlers from `PXEOracleInterface` which would make me very happy because the naming there was terrible (`PXEOracleInterface::getBlockHeader` returned block header of the latest block synchronized by `PXE` while `getTimestamp` and `getBlockNumber` returned values from aztec node.