feat: pxeless playground + embedded wallet
First iteration of the PXEless playground. In order to restore full functionality, it makes use of an `EmbeddedWallet` which besides implementing the `Wallet` interface allows for the creation of accounts from the app.

The databases have been separated into two, a `PlaygroundDB` which contains purely application data and a `WalletDB` that stores accounts and other misc artifacts related to wallet operation.

The `Wallet` interface has been extended with account retrieval, and contract registration an updating has been simplified into a single method.

Co-authored-by: thunkar <gregojquiros@gmail.com>