chore: cycle group cleanup #4 (#16937)
Cleanup and simplification of cycle_group sub-classes
`straus_lookup_table` and `straus_scalar_slice`
- Remove Ultra/Standard conditionals
- Simplify `straus_lookup_table` construction logic that was performing
point additions and x-coordinate checks in two separate loops to avoid
interleaving them for efficiency. We now get this for free since our
circuit is automatically "sorted" by gate type (i.e. via constructing it
in gate-type blocks) so we can combine what was previously two loops
into one. (Note: this does not effect the number of gates but does
effect the order in which they're added which in turn changes the
permutation polys which is what required the vk hash update).
- Remove duplicate native slices computation in `straus_scalar_slice`
(since its also performed as part of the stdlib calculation)