fix: Do not reexecute all pruned blocks in the prune watcher
When deciding whether to slash committee members for an epoch, the epoch
prune watcher tries reexecuting all blocks in the pruned epoch. It also
uses that to decide whether to slash for data withholding, if not all
data is available.

However, the blocks being reexecuted (and the txs being gathered) were
from ALL pruned epochs, which could be more than one if the proof
submission window was long enough. So we ended up slashing committee
members from the first epoch for data withholding offenses perpetrated
by members in future epochs.