refactor: merging PXE interface and its implementation
There will always be only 1 implementation of the PXE interface so it didn't really make sense to have it. In this PR I merged it with PXEService (i.e. copy over the docs) and I rename all the occurrences of PXE Service just as PXE. Now that there is no JSON RPC PXE server that name was misleading.

## Staring into the abyss

Unfortunately, there is a downside to this PR and that is that I needed to drop the proper `PXE` type from `BaseWallet` as Aztec.js cannot import pxe package due to a circular dependency: `pxe --> bbprover --> aztec.js --> pxe`

For this reason now there is the pxe arg defined as `pxe: any`. This will get resolved once `@aztec/wallet-sdk` package is introduced.