feat(avm)!: Handle Empty Calldata
## Empty Calldata

This is a follow-up to #17224 which handles empty calldata within the AVM. The main change is that **we no longer return 0 for empty calldata**:

```rust
   // if calldata.len() == 0 { // <-- removed
   //     0
   //  } else {
        poseidon2_hash_with_separator_slice(calldata, GENERATOR_INDEX__PUBLIC_CALLDATA)
   // }
```
When given an empty array, the AVM will just hash the separator. This is handled inside the circuit by including a special case in the `calldata` trace which indicates we have a `calldata_size` of 0:
```
//  index |  value | context_id | latch |
// -------+--------+------------+-------+
//    0   |    0   |     id     |   1   |
```
This is different from the case where we have calldata of `[ 0 ]`, which would have `index = 1` and `calldata_hash = H([sep, 0])` (with lookups confirming that the poseidon `input_len` is actually 2 and `calldata_size = 1`). I've added some comments to explain that we should never look up the `calldata` trace with an `index` of 0 unless we are looking to confirm that some `context_id` has empty calldata.

### Notes/TODOs

- The protocol will always prepend the function selector to its calldata array, so we won't ever 'see' this case in an e2e test (but we do want to handle it)