refactor: dropping now-unnecessary ArgsHasher
I noticed that we can reuse the functionality I extracted out of `derive_serialize` yesterday in [this PR](https://github.com/AztecProtocol/aztec-packages/pull/17854) also for the hashing of contract function args and return values so I decided to do this PR as well.

Now we have a much better code reuse which is great.