chore: redis-cli and parallel in repo.
This embeds redis-cli and parallel in the repo in `ci3/bin` for use in the GA runner context right at the start of a CI run. These are needed, and installing them via apt can take minutes for some inexplicable reason.
They are only used on the GA runner. Once the ec2 launches we're using whatever is in the AMI / build container.