fix: Reject ECDSA signatures with high-s values and v != 27/28 (#18032)
As defined in EIP2, we do not support ECDSA signatures with s-values in
the upper half of the curve on our contracts:


https://github.com/OpenZeppelin/openzeppelin-contracts/blob/51ab591cd7a47446293a0d5e285792f63cbeb1ea/contracts/utils/cryptography/ECDSA.sol#L176-L184

However, we do allow them when verifying them offchain. This can lead to
a proposer collecting attestations, verifying their signatures
successfully, but then having them rejected when trying to submit a tx.

This PR changes the default behaviour of our Secp256k1 ECDSA utils so
that we reject a signature as invalid if it's in the high-S values. It
also adds a safeguard around web3signer so, if it emits an invalid
signature, it flips it to the correct side.

This PR also makes `recover` to also reject signatures with v-values
other than 27/28, since these are rejected by ECRECOVER. Removing this
check would've meant that an attestor could have supplied a signature
with yParity bit instead of v, which would've been accepted by the node,
posted on chain, and then made the block eligible for invalidation.

Fixes A-182