chore: track snapshot duration
This PR adds new metrics to track snapshot success/errors. This needs to be backported to v2.

I'll open a separate PR to define an alert on the `snapshot_error_count` metric but this one doesn't need to be backported to v2 and I don't want to risk merge conflicts because of it.

Fix A-201