chore: split deploy bootnodes script
Fixes internal issue https://linear.app/aztec-labs/issue/A-195/update-bootnodes-to-use-network-config

Issue text
```
Today, we have a deploy_bootnodes.sh script which performs a number of idempotent actions:

Accepts a list of regions, a machine type, a bucket location, the network name and a few other pieces of config

Creates an ssh key if one doesn't exist and stores it in a GCP secret

Creates a static ip address for each region for the network name

Creates a p2p private key for each region for the network name and stores in GCP secrets

Generates the p2p encoded ENRs

Writes the ENRs to a json file and uploads to the s3 bucket

Launches a VM for each bootnode and runs a startup script, installing dependencies and setting up systemd to keep the bootnode service running. The script pulls the ENRs from the s3 bucket.

The issue is in the location of the store for ENRs (the bucket) as I think it needs to move and also we are not the only ones running bootnodes anymore. The canonical list of bootnodes is in the network config.

What should probably happen is:

The script is split so we do everything other than launch the VMs. The ENRs are then added to the network config.

The VM startup script pulls the list of bootnodes from the network config.
```