chore: ACIR audit - part 3 (#18482)
### 🧾 Audit Context

This is the third PR in the thread of ACIR auditing. It removes code
related to an optimisation for memory constraints that is not affecting
our code. The optimisation aimed at avoiding duplicating range
constraints when a witness `x` was being constrained to be in two ranges
`range_1` and `range_2` with `range_1 < range_2`. Then, the optimisation
ensured the circuit only imposed `x < range_1`.

Noir is already performing this optimisation, see
https://github.com/AztecProtocol/aztec-packages/pull/16112. Given it
doesn't change our circuits, we decided to remove this code so to
simplify acir-related code.

### 🛠️ Changes Made

Removed `minimal_range` and `index_range` from `AcirFormat` and the
relevant methods. Notice that there is no VK change, meaning that this
optimisation was not active.

### ✅ Checklist

- [ ] Audited all methods of the relevant module/class
- [ ] Audited the interface of the module/class with other (relevant)
components
- [ ] Documented existing functionality and any changes made (as per
Doxygen requirements)
- [ ] Resolved and/or closed all issues/TODOs pertaining to the audited
files
- [ ] Confirmed and documented any security or other issues found (if
applicable)
- [ ] Verified that tests cover all critical paths (and added tests if
necessary)
- [ ] Updated audit tracking for the files audited (check the start of
each file you audited)

### 📌 Notes for Reviewers

(Optional) Call out anything that reviewers should pay close attention
to — like logic changes, performance implications, or potential
regressions.

---------

Co-authored-by: Sarkoxed <sarkoxed2013@yandex.ru>