refactor: replacing use of `UniqueNote` with `NoteDao`
Replaced the use of `UniqueNote` with `NoteDao` as having the `UniqueNote` type no longer seemed to make much sense as the only place it was used was the getNotes API on test wallet and PXE and those are considered to be debug-only.

I know this is very low-priority but AI agent is capable of doing such a low-complexity tasks so the cost of doing this here is low.

I moved the NoteDao to stdlib to not force the dependency on PXE in places where it's used.