fix(circuits)!: Ensure global uniqueness of side effect counters
Previously, we were only checking local uniqueness of counters (within a single sideffect type of a single call) to keep gate counts low.
However, some code in the reset circuits assumes global uniqueness of counters, and in general, this is a big source of issues. This PR implements validation for global uniqueness of sideffect counters, by checking in a call request that all counters of all sideffects are unique, that they are within the call sideffect counter bounds, and that they don't fall within the allocated counters for any nested call. This three things ensure that no sideffect counters can repeat in a TX, other than padded side effects that are always injected with MAX_U32_VALUE counter.
The cost of verifying global uniqueness of counters this way is ~5000 gates (~10% of init, ~5% of inner)