refactor!: enshrining owner in notes and state vars
Fixes F-33
Fixes F-179
Fixes https://github.com/AztecProtocol/aztec-packages/issues/15568

In this PR I enshrine owner in notes and in state variables dealing with notes. A motivation for this has been the following:
1. Privacy leak when initializing state variables: https://github.com/AztecProtocol/aztec-packages/issues/15568 (will be tackled in a followup PR by including owner nsk in the nullifier preimage),
2. notes always having a logical owner - we always need the information of an owner to be able to compute a nullifier in a way that the note sender is not able to derive (and hence have "spending" privacy from sender),
3. `PrivateSet` state var has been a bit of a lie as generally we always only worked with a subset of the set - subset of owner's notes (this subset got selected by deriving the storage slot based on the owner address via a `Map`),
4. this also enables stealth addresses as PXE will be able to know the semantic owner and return notes that belong to sub-addresses as well.

I didn't want to tackle https://github.com/AztecProtocol/aztec-packages/issues/15568 in this PR but the nature of the state variables changed and `PrivateMutable` and `PrivateImmutable` became scoped per user which requires them to be intializable for each user. With the original initialization nullifier scheme this didn't work as there was only 1 nullifier per the whole `PrivateMutable`. Introducing `PrivateMutable` and `PrivateImmutable` that can have only 1 owner (the contract admin) is expected to be done in a followup work.

## Issues with this PR
- Lack of these admin state vars make parts of the code quite weird (e.g. account contracts where by definition we have only 1 PublicKeyNote per contract but we access it via `private_immutable.at(self.address).get_note()`).
- State vars naming is now inconsistent: `PublicMutable` contains one value per the whole contract while `PrivateMutable` contains one value per contract-user pair (same goes for `PublicImmutable` and `PrivateImmutable`).