feat(docs): add autogenerated aztec.nr documentation
Adds autogenerated html docs using `nargo doc` to the documentation.

This PR adds the current `next` and `devnet` versions of the html docs. there are updates to the scripts so the aztec-nr docs should be cut as part of the release process as well as for nightly releases.

pretty much all of the additions are the autogenerated html files in `docs/static/aztec-nr-api`, so reviewers can skip these.

I opted to commit so many files because the alternative is to generate them at deploy time, which would add `nargo` as a dependency for building the docs and would increase complexity around generating all of the correct versions