feat: Use debug_tx for processing non-standard rollup propose txs (#18758)
Modified sequencers may post L1 txs to the rollup contract without going
through the expected multicall3. This PR detects if the tx does not
conform to the expected shape, or has any extraneous calls, and if so,
leverages `debug_traceTransaction` or `trace_transaction` for obtaining
the calldata sent to the `propose` tx in the rollup contract.

Adds a new `ETHEREUM_DEBUG_HOSTS` (defaults to `ETHEREUM_HOSTS`) which
is a list of execution client RPCs to use or these debug calls
exclusively.

Nodes now check during startup if any of the ethereum debug hosts
support debug or trace, and if not, will refuse to start unless
`ETHEREUM_ALLOW_NO_DEBUG_HOSTS` is set to true.

Co-authored-by: Phil Windle <philip.windle@gmail.com>