fix(avm_fuzzing): get rid of instruction.argument_tag + more accurate mutations  (#18878)
https://github.com/AztecProtocol/aztec-packages/pull/18857 introduces
`AddressRef` structure, which has its own tag, so the resolved variable
will have the tag of `AddressRef::tag`

So we do not need argument tags for instrutions.


Also with the random tags the most of binary ops are invalid (because
tags for lhs and rhs mismatch). Added mechanism to determenistically
generate almost the same tag for rhs